move people filtering logic into its own file
